requirements.txt
streamlit
pymupdf
python-docx
pytesseract
pillow
rapidfuzz
celery
redis
